+++ src/util.ts
@@ -1,5 +1,6 @@
 import { Collection } from 'mongodb'
 import _ from 'lodash/fp.js'
+import { walkie, Node } from 'obj-walker'
 
 export const setDefaults = (keys: string[], val: any) => {
   const obj: Record<string, any> = {}
@@ -27,3 +28,25 @@ export const omitFieldForUpdate = (omitPaths: string[]) =>
 
 export const getCollectionKey = (collection: Collection) =>
   `${collection.dbName}:${collection.collectionName}`
+
+/**
+ * Remove title and description from a JSON schema.
+ */
+export const removeMetadata = (schema: object) => {
+  const traverse = (x: any) => x.properties || (x.items && { items: x.items })
+  const walkFn = ({ val }: Node) => {
+    if ('title' in val) {
+      delete val.title
+    }
+    if ('description' in val) {
+      delete val.description
+    }
+  }
+  return walkie(schema, walkFn, { traverse })
+}
+
+export function when<T>(condition: any, fn: (x: T) => any) {
+  return function(x: T) {
+    return condition ? fn(x) : x
+  }
+}